+++ micro-services/payments-service/src/index.ts
@@ -1,8 +1,9 @@
 // payments-service entry point
 import 'dotenv/config';
 import express from 'express';
-import webhookRouter from './middlewares/stripeWebhook';
+import webhookRouter from './routes/stripeWebhook';
 import paymentsRouter from './routes/payments';
+import { errorMiddleware } from './middlewares/error.middleware';
 
 const app = express();
 const PORT = process.env.PORT || 3004;
@@ -14,6 +15,8 @@ app.use('/webhook', webhookRouter);
 app.use(express.json());
 
 app.use('/payments', paymentsRouter);
+
+app.use(errorMiddleware);
 
 app.listen(PORT, () => {
   console.log(`payments-service running on port ${PORT}`);
