+++ micro-services/payments-service/src/types/express.d.ts
@@ -1,5 +1,4 @@
 import 'express';
-import { UserRoleType } from '../enum/UserRoleType.enum';
 
 declare module 'express-serve-static-core' {
   interface Request {
@@ -7,7 +6,6 @@ declare module 'express-serve-static-core' {
     user?: {
       userId: string;
       email: string;
-      role?: UserRoleType;
     };
   }
 }
